YUG_INTEGRATION_PACKAGE/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
PySocks==1.7.1
sqlalchemy==2.0.23
pydantic==2.5.0
aiohttp==3.9.1
lxml==4.9.3
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

